+++ src/icons/index.d.ts
@@ -6,11 +6,13 @@ import { BvPlugin, BvComponent } from '../'
 
 // Plugin
 export declare const IconsPlugin: BvPlugin
-
 export declare const BootstrapVueIcons: BvPlugin
 
 // Component: b-icon
 export declare class BIcon extends BvComponent {}
+
+// Component: b-iconstack
+export declare class BIconstack extends BvComponent {}
 
 // Components: b-icon-{icon-name}
 export * from './icons'